pyproject.toml
[project]
name = "agentic-newsletter-aggregator"
version = "0.1.0"
description = "AI-powered newsletter aggregation and summarization system"
authors = [
    {name = "Jan Kneumann"}
]
requires-python = ">=3.11"
dependencies = [
    # Core
    "python-dotenv>=1.0.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",

    # Database
    "sqlalchemy>=2.0.0",
    "alembic>=1.13.0",
    "psycopg2-binary>=2.9.9",

    # Task Queue
    "celery>=5.3.0",
    "redis>=5.0.0",

    # Web Framework
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",

    # Newsletter Ingestion
    "google-auth>=2.27.0",
    "google-auth-oauthlib>=1.2.0",
    "google-auth-httplib2>=0.2.0",
    "google-api-python-client>=2.116.0",
    "feedparser>=6.0.11",
    "beautifulsoup4>=4.12.0",
    "lxml>=5.1.0",

    # Agent Frameworks
    "anthropic>=0.40.0",  # Claude SDK
    "openai>=1.51.0",     # OpenAI SDK

    # HTTP/Requests
    "httpx>=0.26.0",
    "requests>=2.31.0",

    # Email Delivery
    "sendgrid>=6.11.0",

    # Neo4j for Graphiti
    "neo4j>=5.16.0",

    # Utilities
    "python-dateutil>=2.8.2",
    "pytz>=2024.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.1.0",
    "mypy>=1.8.0",
    "ipython>=8.20.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
plugins = ["pydantic.mypy"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"
addopts = "-v --cov=src --cov-report=term-missing"
